added clusters for countries to represent the number of locations where corona virus spreaded across the world
added clusters for countries to represent the number of locations where corona virus spreaded across the world 
This is based on the data that I am getting from the API that I am reaching out to, it gives around 497 locations around the world and then you can click on a location and a table will appear telling the number of cases reported, number of deaths and the number of recovered persons from the clicked location. One thing that the API does not provide is the exact city, for each location it just shows the country and not the exact city name where the pandemic cases were observed. So there can be many clusters for a country. And you can zoom and zoom out each cluster to see specific details